+++ src/open_inwoner/conf/ci.py
@@ -35,7 +35,6 @@ CACHES.update(
         "default": {"BACKEND": "django.core.cache.backends.locmem.LocMemCache"},
         # See: https://github.com/jazzband/django-axes/blob/master/docs/configuration.rst#cache-problems
         "axes": {"BACKEND": "django.core.cache.backends.dummy.DummyCache"},
-        "oidc": {"BACKEND": "django.core.cache.backends.locmem.LocMemCache"},
     }
 )