+++ src/open_inwoner/conf/dev.py
@@ -74,7 +74,6 @@ LOGGING["loggers"].update(
 CACHES = {
     "default": {"BACKEND": "django.core.cache.backends.locmem.LocMemCache"},
     "axes": {"BACKEND": "django.core.cache.backends.dummy.DummyCache"},
-    "oidc": {"BACKEND": "django.core.cache.backends.dummy.DummyCache"},
 }
 
 _MOCK_AUTHENTICATION_BACKENDS = {